mods for wrap-able plasma fractals, added complex gradients
Modifed plasma fractals so that they can be tiled.  The left column is
the same as the right column, and the top is the same as the bottom.

Added CosGradient object to pngsaver that can be used to build
gradients with multiple colors at variable positions.

Added saveCosGradient method that applies a CosGradient to a matrix and
saves it as a png file.

Also removed matrixfix.flatten method.  Numpy.clip is MUCH faster.